Re-organize template functions implementations
